requirements.txt
django>=1.10
djangorestframework>=3.6.3
python-keycloak>=0.8.2
